bp - transfer preparation amountAsDecimal
